backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "appeldoffre-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis[hiredis]>=5.0",
    "celery>=5.3",
    "httpx>=0.27",
    "playwright>=1.44",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "feedparser>=6.0",
    "anthropic>=0.28",
    "openai>=1.30",          # pour les embeddings text-embedding-3-small
    "pgvector>=0.3",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "resend>=2.0",
    "apscheduler>=3.10",
    "pypdf>=4.2",
    "python-docx>=1.1",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx", "ruff", "mypy"]

[tool.setuptools.packages.find]
where = ["."]
